requirements.txt
flask==3.0.2
werkzeug==3.0.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
gdown==4.7.3 